projects/sandbox/mdc/pyproject.toml
[tool.poetry]
name = "mdc"
version = "0.0.1"
description = "Perform local IaaS inference on MDC data with aframe"
authors = ["Alec Gunny <alec.gunny@gmail.com>"]
# readme = "README.md"

[tool.poetry.scripts]
mdc = "mdc.main:main"
deploy-mdc = "mdc.deploy:main"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
ratelimiter = "^1.2"
psutil = "^5.0"

# other ml4gw utilities
typeo = {git = "https://github.com/ML4GW/typeo.git", branch = "main"}

# local aframe utilities
"aframe.analysis" = {path = "../../../libs/analysis", develop = true}
"aframe.deploy" = {path = "../../../libs/deploy", develop = true}
"aframe.io" = {path = "../../../libs/io", develop = true}
"aframe.logging" = {path = "../../../libs/logging", develop = true}

# hermes utilities
"hermes.aeriel" = {path = "../../../hermes/hermes/hermes.aeriel", extras=["serve"], develop = true}
"hermes.stillwater" = {path = "../../../hermes/hermes/hermes.stillwater", develop = true}

[tool.poetry.dev-dependencies]
pytest = "^7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
